PTP Dual nic support implmentation. This PR adds support to handle dual supporbouandary clock. One of the nic will be running PTP without phc2sys. The event framework handles various events based on logs from both nic ptp4l and single phc2sys.
